Added ODE-solver version of the mass-spring-damper environment
Added a version of the mass-spring-damper environment that uses a true ODE solution rather than a simple approximate. As of now, it does not work with `grad_method=GradMethods.AUTO_DIFF` but does with `grad_method=GradMethods.FINITE_DIFF` in the MPC call. The naive finite differences solution is *sloooow*.